src/server/api/internal/uploadDocument/uploadDocuments.spec.ts
// import httpMocks from 'node-mocks-http';
// import { createSuccessResponse, createErrorResponse } from '@tamm/response';
// import S3 from 'server/services/s3';
// import * as upload from './uploadDocuments';

// const fileType = require('file-type');

// jest.mock('server/services/s3');
// jest.mock('@tamm/response');
// jest.mock('read-chunk');
// jest.mock('file-type', () => jest.fn());

// describe('uploadDocuments', () => {
//   let req: httpMocks.MockRequest<any>;
//   let res: httpMocks.MockResponse<any>;
//   beforeEach(() => {
//     req = httpMocks.createRequest();
//     res = httpMocks.createResponse();
//     req = {
//       session: {
//         id: 'test-id',
//       },
//       uploaded: {
//         filePath: 'file/path',
//         newFileName: 'new_name',
//       },
//       file: true,
//     };
//   });

//   afterEach(() => {
//     jest.clearAllMocks();
//   });

//   it('should give error response if no file in req', async () => {
//     req.file = false;
//     await upload.uploadDocuments(req, res);
//     expect(createErrorResponse).toHaveBeenCalledWith(
//       req,
//       res,
//       500,
//       'File not found',
//       {},
//     );
//   });
//   it('should give error response if no orginal name', async () => {
//     req.uploaded.orginalName = 'some_name';
//     req.file = false;
//     await upload.uploadDocuments(req, res);
//     expect(createErrorResponse).toHaveBeenCalledWith(
//       req,
//       res,
//       500,
//       'File not found',
//       {},
//     );
//   });
//   it('should give error response if file is not pdf', async () => {
//     S3.uploadDocument = jest.fn().mockRejectedValue('error');
//     await upload.uploadDocuments(req, res);
//     expect(createErrorResponse).toHaveBeenCalledWith(
//       req,
//       res,
//       'Only upload PDF/JPG files',
//       {},
//     );
//   });

//   it('should give success response for pdf', async () => {
//     fileType.mockImplementation(() => ({
//       ext: 'pdf',
//       mime: 'application/pdf',
//     }));

//     const response = {
//       data: {
//         status: 'success',
//         data: {
//           fileName: 'test_name',
//           message: 'message',
//         },
//       },
//     };
//     S3.uploadDocument = jest
//       .fn()
//       .mockResolvedValueOnce(Promise.resolve(response));
//     await upload.uploadDocuments(req, res);
//     expect(createSuccessResponse).toHaveBeenCalled();
//   });

//   it('should give error response for failed response', async () => {
//     fileType.mockImplementation(() => ({
//       ext: 'pdf',
//       mime: 'pdf',
//     }));
//     const response = {
//       data: {
//         status: 'failed',
//         data: {},
//       },
//     };
//     S3.uploadDocument = jest
//       .fn()
//       .mockResolvedValueOnce(Promise.resolve(response));
//     await upload.uploadDocuments(req, res);
//     expect(createErrorResponse).toHaveBeenCalled();
//   });

//   it('should give error response', async () => {
//     S3.uploadDocument = jest.fn().mockRejectedValue('error');
//     await upload.uploadDocuments(req, res);
//     expect(createErrorResponse).toHaveBeenCalled();
//   });
// });
it('test', () => {
  expect(1).toBe(1);
});
